moved sub logic over to common and fixed some dates/comments
